Move get functionality out of AccumuloIndex to IndexTableWrapper
